Refactor folder structure for components under `src/view`.
Moved all files from `src/pages` to `src/view` for improved organization and consistency. Updated imports across the application to reflect the new file paths. Additionally, adjusted `package-lock.json` with changes for dependency updates or formatting consistency.
