+++ src/view/home/styles.ts
