+++ src/view/menu-list/styles.ts
