+++ src/view/onboarding/index.tsx
