+++ src/view/onboarding/styles.ts
