+++ src/view/product-registration/index.tsx
