+++ src/view/product-registration/styles.ts
