+++ src/view/restaurant-registration/index.tsx
