+++ src/view/restaurant-registration/styles.ts
