+++ src/view/user-registration/index.tsx
