+++ src/view/user-registration/styles.ts
